refactor(dns): make inline modules
