--- transports/dns/src/config.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "async-std")]
-pub mod async_std;
-
-#[cfg(feature = "tokio")]
-pub mod tokio;
