+++ transports/dns/src/lib.rs
@@ -57,17 +57,77 @@
 
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-mod config;
-
 #[cfg(feature = "async-std")]
-pub use config::async_std;
+pub mod async_std {
+    use async_std_resolver::AsyncStdResolver;
+    use parking_lot::Mutex;
+    use std::{io, sync::Arc};
+    use trust_dns_resolver::{
+        config::{ResolverConfig, ResolverOpts},
+        system_conf,
+    };
+
+    /// A `Transport` wrapper for performing DNS lookups when dialing `Multiaddr`esses
+    /// using `async-std` for all async I/O.
+    pub type Config<T> = crate::Config<T, AsyncStdResolver>;
+
+    impl<T> Config<T> {
+        /// Creates a new [`Config`] from the OS's DNS configuration and defaults.
+        pub async fn system(inner: T) -> Result<Config<T>, io::Error> {
+            let (cfg, opts) = system_conf::read_system_conf()?;
+            Self::custom(inner, cfg, opts).await
+        }
+
+        /// Creates a [`Config`] with a custom resolver configuration and options.
+        pub async fn custom(
+            inner: T,
+            cfg: ResolverConfig,
+            opts: ResolverOpts,
+        ) -> Result<Config<T>, io::Error> {
+            Ok(Config {
+                inner: Arc::new(Mutex::new(inner)),
+                resolver: async_std_resolver::resolver(cfg, opts).await,
+            })
+        }
+    }
+}
 
 #[cfg(feature = "async-std")]
 #[deprecated(note = "Use `async_std::Config` instead.")]
 pub type DnsConfig<T> = async_std::Config<T>;
 
 #[cfg(feature = "tokio")]
-pub use config::tokio;
+pub mod tokio {
+    use parking_lot::Mutex;
+    use std::sync::Arc;
+    use trust_dns_resolver::{system_conf, TokioAsyncResolver};
+
+    /// A `Transport` wrapper for performing DNS lookups when dialing `Multiaddr`esses
+    /// using `tokio` for all async I/O.
+    pub type Config<T> = crate::Config<T, TokioAsyncResolver>;
+
+    impl<T> Config<T> {
+        /// Creates a new [`Config`] from the OS's DNS configuration and defaults.
+        pub fn system(inner: T) -> Result<crate::Config<T, TokioAsyncResolver>, std::io::Error> {
+            let (cfg, opts) = system_conf::read_system_conf()?;
+            Self::custom(inner, cfg, opts)
+        }
+
+        /// Creates a [`Config`] with a custom resolver configuration
+        /// and options.
+        pub fn custom(
+            inner: T,
+            cfg: trust_dns_resolver::config::ResolverConfig,
+            opts: trust_dns_resolver::config::ResolverOpts,
+        ) -> Result<crate::Config<T, TokioAsyncResolver>, std::io::Error> {
+            // TODO: Make infallible in next breaking release. Or deprecation?
+            Ok(Config {
+                inner: Arc::new(Mutex::new(inner)),
+                resolver: TokioAsyncResolver::tokio(cfg, opts),
+            })
+        }
+    }
+}
 
 #[cfg(feature = "tokio")]
 #[deprecated(note = "Use `tokio::Config` instead.")]
